pyproject.toml
[project]
name = "LogTT"
version = "0.1.0"
description = "结构化日志分析与可视化系统"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "loguru>=0.7.3",
    "pandas>=2.3.3",
    "pyside6-fluent-widgets[full]>=1.11.0",
    "pyyaml>=6.0.3",
    "regex>=2025.11.3",
]